Added abstract class DrawerActivity with logic for activity with drawer Change template activity for work with DrawerActivity
