--- app/src/main/res/layout/test_widget.xml
@@ -1,26 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id= "@+id/test_widget"
-
-
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#09C"
-    android:padding="@dimen/widget_margin">
-
-    <TextView
-        android:id="@+id/appwidget_text"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_centerInParent="true"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:background="#09C"
-        android:contentDescription="@string/appwidget_textEx"
-        android:text="@string/appwidget_textEx"
-        android:textColor="#000000"
-        android:textSize="20sp"
-        android:textStyle="" />
-
-</RelativeLayout>
