+++ app/src/main/res/layout/x_template_empty_layout.xml
@@ -19,10 +19,10 @@
             app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
 
         <TextView
-            android:id="@+id/template_textView"
+            android:id="@+id/template_text_view"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/error_template"
+            android:text="@string/x_template_empty_activity_template"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
